update readme rename ui to components remove empty folders (as they are already documented on the readme)
