+++ src/components/App.js
@@ -1,6 +1,6 @@
 import React from 'react';
-import 'ui/App.css';
-import CommonComponent from 'ui/common/CommonComponent';
+import 'components/App.css';
+import CommonComponent from 'components/common/CommonComponent';
 
 function App() {
   return (
