+++ src/components/__tests__/App.test.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
-import App from 'ui/App';
+import App from 'components/App';
 
 test('renders without crashing', () => {
   const div = document.createElement('div');
