+++ src/components/common/__tests__/CommonComponent.test.js
@@ -1,7 +1,7 @@
 import '@testing-library/jest-dom/extend-expect';
 import React from 'react';
 import { render } from '@testing-library/react';
-import CommonComponent from 'ui/common/CommonComponent';
+import CommonComponent from 'components/common/CommonComponent';
 
 test('renders a message', () => {
   const { container, getByText } = render(<CommonComponent />);
